+++ src/briefy/leica/models/billing_info/__init__.py
@@ -3,7 +3,6 @@ from briefy.leica.db import Base
 from briefy.leica.models import mixins
 from briefy.leica.vocabularies import TaxIdTypes
 from sqlalchemy.ext.declarative import declared_attr
-from sqlalchemy.ext.hybrid import hybrid_property
 
 import colander
 import sqlalchemy as sa
@@ -123,37 +122,3 @@ class BillingInfo(TaxInfo, mixins.BillingAddress, mixins.LeicaVersionedMixin, Ba
         if cls_name == 'billinginfo':
             args['polymorphic_on'] = cls.type
         return args
-
-    @hybrid_property
-    def title(self):
-        """Return fullname as title."""
-        return self.first_name + ' ' + self.last_name
-
-    @title.setter
-    def title(self, value: str):
-        """Avoid set title since it is a computed field."""
-        return ValueError(
-            'You can not set the billing info title. Please update first_name and last_name.'
-        )
-
-    @title.expression
-    def title(cls):
-        """Return fullname as title."""
-        return cls.first_name + ' ' + cls.last_name
-
-    @sautils.observes('first_name', 'last_name')
-    def _title_observer(self, first_name, last_name):
-        """Calculate dates on a change of a state."""
-        self._title = first_name + ' ' + last_name
-
-    def to_dict(self, excludes: list=None, includes: list=None) -> dict:
-        """Return a dictionary with fields and values used by this Class.
-
-        :param excludes: attributes to exclude from dict representation.
-        :param includes: attributes to include from dict representation.
-        :returns: Dictionary with fields and values used by this Class
-        """
-        data = super().to_dict(excludes, includes)
-        data['slug'] = self.slug
-        data['title'] = self.title
-        return data