+++ src/briefy/leica/models/project/__init__.py
@@ -1,11 +1,11 @@
 """Briefy Leica Project model."""
+from briefy.common.db.models import Item
 from briefy.common.utils import schema
 from briefy.common.utils.data import Objectify
 from briefy.common.vocabularies.categories import CategoryChoices
 from briefy.common.vocabularies.roles import Groups
 from briefy.leica.cache import cache_region
 from briefy.leica.cache import enable_cache
-from briefy.leica.db import Item
 from briefy.leica.models import mixins
 from briefy.leica.models.project import workflows
 from briefy.leica.utils.user import add_user_info_to_state_history
@@ -524,7 +524,6 @@ class Project(CommercialInfoMixin, mixins.ProjectRolesMixin,
         excludes = list(excludes) if excludes else []
         excludes.append('orders')
         data = super().to_dict(excludes=excludes, includes=includes)
-        data['slug'] = self.slug
         data['price'] = self.price
         data['category'] = self.category.value \
             if isinstance(self.category, CategoryChoices) else self.category