+++ tests/models/professional/test_professional.py
@@ -3,7 +3,6 @@ from briefy.leica import models
 from conftest import BaseModelTest
 
 import pytest
-import transaction
 
 
 @pytest.mark.usefixtures('create_dependencies')
@@ -27,6 +26,19 @@ class TestProfessionalModel(BaseModelTest):
             web_request.user = role
             wf.context = role
         return obj, wf, web_request
+
+    def test_add_pool_to_professional(self, instance_obj):
+        """Add job pools to the professional."""
+        pools = models.Pool.query().all()
+        assert len(pools) == 3
+        assert len(instance_obj.pools) == 0
+
+        for item in pools:
+            instance_obj.pools.append(item)
+            assert instance_obj in item.professionals
+            assert len(item.professionals) == 1
+
+        assert len(instance_obj.pools) == 3
 
     @pytest.mark.parametrize('origin_state', ['pending'])
     @pytest.mark.parametrize('role_name', ['qa', 'scout'])
@@ -190,16 +202,3 @@ class TestProfessionalModel(BaseModelTest):
         wf.assign(fields={'pools_ids': [pool_id, ]})
         assert obj.state == origin_state
         assert obj.state_history[-1]['transition'] == 'assign'
-
-    def test_add_pool_to_professional(self, instance_obj):
-        """Add job pools to the professional."""
-        pools = models.Pool.query().all()
-        assert len(pools) == 3
-        assert len(instance_obj.pools) == 0
-
-        for item in pools:
-            instance_obj.pools.append(item)
-            assert instance_obj in item.professionals
-            assert len(item.professionals) == 1
-
-        assert len(instance_obj.pools) == 3